:ok_hand: Passing Icon node to IconTab component instead of mapping over it inside the IconTab component
